Fix issue with hash not being stored in session (#5)
There is an issue with the current version whereby if the UserVisit is saved,
but the request hash is not stored in the session, every subsequent request
will raise an IntegrityError - as it will try and save another UserVisit.
I haven't got to the bottom of why some hashes are not being store properly,
but it causes a cascade of warnings. This should fix the cascade, by ensuring
that the hash is stored even in the event of an IntegrityError.