Add context JSONField (#17)
This PR adds a JSONField to the UserVisit model that can be used to
store ad hoc data against a request. The canonical example for this is
GeoIP data - see https://github.com/yunojuno/django-user-visit/issues/15,
but it can be used for anything (so long as it is JSON serializable).